Correct Debian/Ubuntu package list
- mktemp and tr are part of package coreutils
- find is part of package findutils

Note: both findutils and coreutils are marked as essential, so should
always be installed. So listing these packages is not strictly necessary,
but it also doesn't really hurt.

Sync the package lists between the README.md and utils.py. Switch to fuse3
in utils.py, to match the description in the README.