feat: generic prompts
Works on all three UIs offers a generic function to ask a question that platform independent.
If the user fails to offer a response, the installer will terminate.

In the GUI this still works, however it may not be desirable to prompt the user for each question.
So long as we don't attempt to access the variable before the user has had a chance to put in their preferences it will not prompt them
Changed the GUI to gray out the other widgets if the product is not selected.
start_ensure_config is called AFTER product is set, if it's called before it attempts to figure out which platform it's on, prompting the user with an additional dialog (not ideal, but acceptable)