fix: initial install choice selection
Setting switch_q to 1 when contents were changed removed the pending ask screen, causing install to stop halfway

Also it's possible for the TUI to ask for the installed_faithlife_product_release before install_dir is set, causing two threads to both be asking, causing a invalid response. The application would probably recover, but it's not ideal.