fix: only remove unused updateURLs
When deleting the application update, there are entries added to the UpdateUrls table which should be cleared to prevent bloat,
however when normal resources are downloaded the UpdateUrls table is also updated.

I failed to test downloading new resources with the original code.

Tested:
- v38.1 installation hitting the download updates banner. Downloads the book updates, but not the application update
- v38.1 install downloading new books from the library pane
- clean v39.1 install downloading resources. Both initially and additional in the library tab after download

and also fixed a bug where if the process kill for logos stop fails, it wouldn't set logos state properly

Fixes: #275 #262